standardize Parameter class
