+++ src/qua_dashboards/core/__init__.py
@@ -1,7 +1,9 @@
 from .base_component import *  # noqa: F403
 from .base_updatable_component import *  # noqa: F403
+from .basic_parameter import *  # noqa: F403
 
 __all__ = [
     *base_component.__all__,  # type: ignore  # noqa: F405
     *base_updatable_component.__all__,  # noqa: F405
+    *basic_parameter.__all__,  # noqa: F405
 ]
